Sanitize build system.
CRC computation moved from build_payload.py to stage0_unsafemode.py

stage0_unsafemode.py now called from stage1
