+++ .gitignore
@@ -57,3 +57,8 @@ dkms.conf
 build/
 
 *.bin
+*.dat
+*.cia
+
+*.map
+*.lst
