+++ bb3_installer/payload/Makefile
@@ -1,15 +1,28 @@
-include $(DEVKITARM)/base_rules
+#---------------------------------------------------------------------------------
+.SUFFIXES:
+#---------------------------------------------------------------------------------
+
+ifeq ($(strip $(DEVKITARM)),)
+$(error "Please set DEVKITARM in your environment. export DEVKITARM=<path to>devkitARM")
+endif
+
+TOPDIR ?= $(CURDIR)
+include $(DEVKITARM)/base_tools
 
 HAXNAME := rop_payload
+
+.PHONY: all clean
 
 all:	$(HAXNAME).bin
 
 clean:
-	rm -f $(HAXNAME).elf $(HAXNAME).bin
+	@rm -f $(HAXNAME).elf $(HAXNAME).bin
+	@echo clean ...
 
 $(HAXNAME).bin: $(HAXNAME).elf
-	arm-none-eabi-objcopy -O binary $(HAXNAME).elf $(HAXNAME).bin
+	$(OBJCOPY) -O binary $(HAXNAME).elf $(HAXNAME).bin
+	@echo built ... $(notdir $@)
 
 $(HAXNAME).elf:	$(HAXNAME).s
-	arm-none-eabi-gcc -x assembler-with-cpp -nostartfiles -nostdlib -Ttext=0x00686000 $< -o $(HAXNAME).elf
+	$(CC) -x assembler-with-cpp -nostartfiles -nostdlib -Ttext=0x00686000 $< -o $(HAXNAME).elf
