--- mini_b9s_installer/build.bat
@@ -1,3 +1,0 @@
-rem bin2c -d b9s.h -o b9s.c boot9strap
-make
-pause
