+++ mini_b9s_installer/linker.specs
@@ -1,0 +1,4 @@
+%rename link                old_link
+
+*link:
+%(old_link) -T %:getenv(TOPDIR /linker.ld) --nmagic --gc-sections
