+++ slotTool/resources/cia.rsf
@@ -4,7 +4,7 @@ BasicInfo:
   Logo                    : Homebrew
 
 RomFs:
-  RootPath                : $(APP_ROMFS)
+#  RootPath                : $(APP_ROMFS)
 
 TitleInfo:
   Category                : Application
