+++ stage1/Makefile
@@ -1,3 +1,16 @@
+#---------------------------------------------------------------------------------
+.SUFFIXES:
+#---------------------------------------------------------------------------------
+
+ifeq ($(strip $(DEVKITARM)),)
+$(error "Please set DEVKITARM in your environment. export DEVKITARM=<path to>devkitARM")
+endif
+
+TOPDIR ?= $(CURDIR)
+include $(DEVKITARM)/base_tools
+
+UMSDIR		:=	$(CURDIR)/..
+
 REGION := USA
 ifeq ($(strip $(REGION)),)
 $(error "Set the REGION Makefile parameter.")
@@ -19,14 +32,22 @@ ifeq ($(REGCODE),00)
 	$(error "Unsupported region.")
 endif
 
-all:	rop_payload.bin
+.PHONY: all clean
+
+all:	slot1.bin
 
 clean:
-	rm -f $(HAXNAME).elf rop_payload.bin
+	@rm -f $(HAXNAME).elf rop_payload.bin slot1.bin
+	@echo clean ...
 
-rop_payload.bin: $(HAXNAME).elf
-	arm-none-eabi-objcopy -O binary $(HAXNAME).elf rop_payload.bin
+slot1.bin: $(HAXNAME).bin $(UMSDIR)/stage0_unsafemode.py
+	@python $(UMSDIR)/stage0_unsafemode.py
+	@echo built ... $(notdir $@)
+
+$(HAXNAME).bin: $(HAXNAME).elf
+	$(OBJCOPY) -O binary $(HAXNAME).elf rop_payload.bin
+	@echo built ... $(notdir $@)
 
 $(HAXNAME).elf:	$(HAXNAME).s
-	arm-none-eabi-gcc -x assembler-with-cpp -nostartfiles -nostdlib -Ttext=0x00276000 $< -o $(HAXNAME).elf
+	$(CC) -x assembler-with-cpp -nostartfiles -nostdlib -Ttext=0x00276000 $< -o $(HAXNAME).elf
